Fix duplicate items from fetching into redux store
